[TreeView] Standardize folder structure with index
