+++ src/components/TreeView/index.ts
@@ -1,0 +1,3 @@
+import TreeView from "components/TreeView/TreeView";
+
+export default TreeView;
